+++ Package.swift
@@ -11,17 +11,20 @@ let package = Package(
     products: [
         .library(name: "SpotifyWebAPIInterchangeKit",
                  targets: ["SpotifyWebAPIInterchangeKit"]),
+        .library(name: "SpotifyWebAPIDataPresets",
+                 targets: ["SpotifyWebAPIDataPresets"])
     ],
     dependencies: [
-        .package(url: "https://github.com/antarianLogic/interchange", from: "1.0.4"),
-        .package(url: "https://github.com/antarianLogic/json-utils", from: "1.0.0")
+        .package(url: "https://github.com/antarianLogic/interchange", from: "1.0.4")
     ],
     targets: [
         .target(name: "SpotifyWebAPIInterchangeKit",
-                dependencies: [.product(name: "Interchange", package: "interchange"),
-                               .product(name: "JSONUtils", package: "json-utils")],
-                resources: [.copy("DataPresets/JSON")]),
+                dependencies: [.product(name: "Interchange", package: "interchange")]),
+        .target(name: "SpotifyWebAPIDataPresets",
+                dependencies: ["SpotifyWebAPIInterchangeKit"],
+                resources: [.copy("JSON")]),
         .testTarget(name: "SpotifyWebAPIInterchangeKitTests",
-                    dependencies: ["SpotifyWebAPIInterchangeKit"]),
+                    dependencies: ["SpotifyWebAPIInterchangeKit",
+                                   "SpotifyWebAPIDataPresets"]),
     ]
 )