+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2026-02-04
+
+### Added
+- `JSONDecoderHelpers.swift`, absorbed code from dependency on private repo that has been removed
+
+### Changed
+- Moved data presets into their own product separate from the main SpotifyWebAPIInterchangeKit product
+
+### Removed
+- dependency on private repo
+
 ## [1.0.4] - 2026-02-02
 
 ### Added